weather-app/backend/requirements.txt
fastapi
uvicorn
langchain
langchain-openai
python-dotenv
httpx
requests
